% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_weight_decay.R
\name{add_weight_decay}
\alias{add_weight_decay}
\title{Add weight decay to any autoencoder}
\usage{
add_weight_decay(learner, decay = 0.02)
}
\arguments{
\item{learner}{The \code{"ruta_autoencoder"} object}

\item{decay}{Numeric value indicating the amount of decay}
}
\value{
An autoencoder object which contains the weight decay
}
\description{
Adds a weight decay regularization to the encoding layer of a
given autoencoder
}
