% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/loss.R
\name{as_loss}
\alias{as_loss}
\alias{as_loss.character}
\alias{as_loss.ruta_loss}
\title{Coercion to ruta_loss}
\usage{
as_loss(x)

\method{as_loss}{character}(x)

\method{as_loss}{ruta_loss}(x)
}
\arguments{
\item{x}{Object to be converted into a loss}
}
\value{
A \code{"ruta_loss"} construct
}
\description{
Generic function to coerce objects into loss objects.
}
