% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/layers.R, R/network.R
\name{as_network}
\alias{as_network}
\alias{as_network.ruta_layer}
\alias{as_network.ruta_network}
\alias{as_network.numeric}
\alias{as_network.integer}
\title{Coercion to ruta_network}
\usage{
as_network(x)

\method{as_network}{ruta_layer}(x)

\method{as_network}{ruta_network}(x)

\method{as_network}{numeric}(x)

\method{as_network}{integer}(x)
}
\arguments{
\item{x}{Object to be converted into a network}
}
\value{
A \code{"ruta_network"} construct
}
\description{
Generic function to coerce objects into networks.
}
\examples{
net <- as_network(c(784, 1000, 32))
}
