% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{autoencode}
\alias{autoencode}
\title{Automatically compute an encoding of a data matrix}
\usage{
autoencode(data, encoding_dim, activation = "linear")
}
\arguments{
\item{data}{Numeric matrix to be encoded}

\item{encoding_dim}{Number of variables to be used in the encoding}
}
\value{
Matrix containing the encodings
}
\description{
Trains an autoencoder adapted to the data and extracts its encoding for the
same data matrix.
}
\examples{
\dontrun{
inputs <- as.matrix(iris[, 1:4])
# Train a basic autoencoder and generate a 2-variable encoding
encoded <- autoencode(inputs, 2)
}

}
\seealso{
\code{\link{autoencoder}}
}
