% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.r
\name{autoencoder}
\alias{autoencoder}
\title{Create an autoencoder learner}
\usage{
autoencoder(layers, sparse = F, contractive = F)
}
\arguments{
\item{layers}{Layer construct of class \code{"rutaNetwork"}}

\item{sparse}{Logical}

\item{contractive}{Logical}
}
\description{
Create an autoencoder learner
}
