% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_contractive.R
\name{autoencoder_contractive}
\alias{autoencoder_contractive}
\title{Create a contractive autoencoder}
\usage{
autoencoder_contractive(network, loss, weight)
}
\arguments{
\item{network}{Layer construct of class \code{"ruta_network"}}

\item{loss}{Character string specifying the reconstruction error part of the loss function}

\item{weight}{Weight assigned to the contractive loss}
}
\value{
A construct of class \code{"ruta_autoencoder"}
}
\description{
A contractive autoencoder adds a penalty term to the loss
function of a basic autoencoder which attempts to induce a contraction of
data in the latent space.
}
