% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_denoising.R
\name{autoencoder_denoising}
\alias{autoencoder_denoising}
\title{Create a denoising autoencoder}
\usage{
autoencoder_denoising(network, loss, ratio = 0.05)
}
\arguments{
\item{network}{Layer construct of class \code{"ruta_network"}}

\item{loss}{Loss function to be optimized}

\item{ratio}{The ratio of each instance in the input data which will be
altered by random noise}
}
\value{
A construct of class \code{"ruta_autoencoder"}
}
\description{
A denoising autoencoder trains with noisy data in order to
create a model able to reduce noise in reconstructions from input data
}
\references{
\itemize{
\item \href{https://dl.acm.org/citation.cfm?id=1390294}{Extracting and composing robust features with denoising autoencoders}
}
}
\seealso{
Other autoencoder variants: \code{\link{autoencoder_contractive}},
  \code{\link{autoencoder_robust}},
  \code{\link{autoencoder_sparse}},
  \code{\link{autoencoder_variational}},
  \code{\link{autoencoder}}
}
