% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_robust.R
\name{autoencoder_robust}
\alias{autoencoder_robust}
\title{Create a robust autoencoder}
\usage{
autoencoder_robust(network, sigma)
}
\arguments{
\item{network}{Layer construct of class \code{"ruta_network"}}

\item{sigma}{Sigma parameter in the kernel used for correntropy}
}
\value{
A construct of class \code{"ruta_autoencoder"}
}
\description{
A robust autoencoder uses a special objective function,
correntropy, a localized similarity measure which makes it less sensitive
to noise in data.
}
