% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_sparse.R
\name{autoencoder_sparse}
\alias{autoencoder_sparse}
\title{Sparse autoencoder}
\usage{
autoencoder_sparse(network, loss, high_probability = 0.1, weight = 0.2)
}
\arguments{
\item{network}{Layer construct of class \code{"ruta_network"}}

\item{loss}{Character string specifying a loss function}

\item{high_probability}{Expected probability of the high value of the
encoding layer. Set this to a value near zero in order to minimize
activations in that layer.}

\item{weight}{The weight of the sparsity regularization}
}
\value{
A construct of class \code{"ruta_autoencoder"}
}
\description{
Creates a representation of a sparse autoencoder.
}
\seealso{
\code{\link{sparsity}}, \code{\link{make_sparse}}
}
