% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_variational.R
\name{autoencoder_variational}
\alias{autoencoder_variational}
\title{Build a variational autoencoder}
\usage{
autoencoder_variational(intermediate, latent_dim,
  loss = "binary_crossentropy")
}
\arguments{
\item{intermediate}{Intermediate size}

\item{latent_dim}{Latent space dimension}

\item{loss}{Reconstruction error to be combined with KL divergence in order to compute
the variational loss}
}
\description{
Build a variational autoencoder
}
