% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{configure}
\alias{configure}
\title{Configure a learner object with the associated Keras objects}
\usage{
configure(learner, input_shape)
}
\arguments{
\item{learner}{A \code{"ruta_autoencoder"} object}

\item{input_shape}{An integer describing the number of variables in the input data}
}
\value{
Same learner passed as parameter, with new members \code{input_shape},
\code{models} and \code{keras_loss}
}
\description{
This function builds all the necessary Keras objects in order
to manipulate internals of the models or train them. Users won't generally
need to call this function, since \code{train} will do so automatically.
}
