% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_contractive.R
\name{contraction}
\alias{contraction}
\title{Contractive loss}
\usage{
contraction(rec_err, weight)
}
\arguments{
\item{rec_err}{Original reconstruction error to be combined with the
contractive loss}

\item{weight}{Weight assigned to the contractive loss}
}
\value{
A loss object which can be converted into a Keras loss
}
\description{
This is a wrapper for a loss which induces a contraction in the
latent space.
}
\seealso{
\code{\link{autoencoder_contractive}}

Other loss functions: \code{\link{correntropy}},
  \code{\link{variational_loss}}
}
