% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_robust.R
\name{correntropy}
\alias{correntropy}
\title{Correntropy loss}
\usage{
correntropy(sigma)
}
\arguments{
\item{sigma}{Sigma parameter in the kernel}
}
\value{
A \code{"ruta_loss"} object
}
\description{
A wrapper for the correntropy loss function
}
\seealso{
\code{\link{autoencoder_robust}}

Other loss functions: \code{\link{contraction}},
  \code{\link{loss_variational}}
}
