% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.r
\name{decode}
\alias{decode}
\title{Retrieve decoding of encoded data}
\usage{
decode(ae, data)
}
\arguments{
\item{data}{data.frame to be encoded}

\item{model}{Autoencoder model}
}
\description{
Extracts the decodification calculated by a trained autoencoder for the specified
data.
}
