% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.r
\name{dense}
\alias{dense}
\title{Create a fully-connected neural layer}
\usage{
dense(units, activation = "linear")
}
\arguments{
\item{units}{Number of units}

\item{activation}{Optional, string indicating activation function (linear by default)}
}
\value{
A construct with class \code{"ruta_layer"}
}
\description{
Wrapper for a dense/fully-connected layer.
}
\examples{
dense(30, "tanh")
}
\seealso{
Other neural layers: \code{\link{input}},
  \code{\link{output}}
}
