% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{dropout}
\alias{dropout}
\title{Dropout layer}
\usage{
dropout(rate = 0.5)
}
\arguments{
\item{rate}{The fraction of affected units}
}
\value{
A construct of class \code{"ruta_network"}
}
\description{
Randomly sets a fraction \code{rate} of input units to 0 at each update during training
time, which helps prevent overfitting.
}
\seealso{
Other neural layers: 
\code{\link{conv}()},
\code{\link{deconv}()},
\code{\link{dense}()},
\code{\link{input}()},
\code{\link{layer_keras}()},
\code{\link{output}()},
\code{\link{variational_block}()}
}
\concept{neural layers}
