% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.r
\name{encode}
\alias{encode}
\title{Retrieve encoding of data}
\usage{
encode(ae, data)
}
\arguments{
\item{ae}{Autoencoder model}

\item{data}{data.frame to be encoded}
}
\description{
Extracts the encoding calculated by a trained autoencoder for the specified
data.
}
