% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.r
\name{encoding_index}
\alias{encoding_index}
\title{Get the index of the encoding}
\usage{
encoding_index(net)
}
\arguments{
\item{net}{A network of class \code{"ruta_network"}}
}
\value{
Index of the middle layer
}
\description{
Calculates the index of the middle layer of an encoder-decoder network.
}
