% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_denoising.R, R/filter.R
\name{apply_filter.ruta_noise}
\alias{apply_filter.ruta_noise}
\alias{apply_filter}
\alias{apply_filter.ruta_custom}
\title{Apply filters}
\usage{
\method{apply_filter}{ruta_noise}(filter, data, ...)

apply_filter(filter, data, ...)

\method{apply_filter}{ruta_custom}(filter, data, ...)
}
\arguments{
\item{filter}{Filter object to be applied}

\item{data}{Input data to be filtered}

\item{...}{Other parameters}
}
\description{
Apply a filter to input data, generally a noise filter in
order to train a denoising autoencoder. Users won't generally need to use
these functions
}
\seealso{
\code{\link{autoencoder_denoising}}
}
