% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{input}
\alias{input}
\title{Create an input layer}
\usage{
input()
}
\value{
A construct with class \code{"ruta_network"}
}
\description{
This layer acts as a placeholder for input data. The number of units is not
needed as it is deduced from the data during training.
}
\seealso{
Other neural layers: \code{\link{dense}},
  \code{\link{dropout}}, \code{\link{layer_keras}},
  \code{\link{output}}, \code{\link{variational_block}}
}
\concept{neural layers}
