% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_denoising.R
\name{is_denoising}
\alias{is_denoising}
\title{Detect whether an autoencoder is denoising}
\usage{
is_denoising(learner)
}
\arguments{
\item{learner}{A \code{"ruta_autoencoder"} object}
}
\value{
Logical value indicating if a noise generator was found
}
\description{
Detect whether an autoencoder is denoising
}
\seealso{
\verb{\link{noise}}, \verb{\link{autoencoder_denoising}}, \verb{\link{make_denoising}}
}
