% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_orthonormal.R
\name{is_orthonormal}
\alias{is_orthonormal}
\title{Detect whether an autoencoder is contractive}
\usage{
is_orthonormal(learner)
}
\arguments{
\item{learner}{A \code{"ruta_autoencoder"} object}
}
\value{
Logical value indicating if a contractive loss was found
}
\description{
Detect whether an autoencoder is contractive
}
\seealso{
\verb{\link{contraction}}, \verb{\link{autoencoder_contractive}}, \verb{\link{make_contractive}}
}
