% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_robust.R
\name{is_robust}
\alias{is_robust}
\title{Detect whether an autoencoder is robust}
\usage{
is_robust(learner)
}
\arguments{
\item{learner}{A \code{"ruta_autoencoder"} object}
}
\value{
Logical value indicating if a correntropy loss was found
}
\description{
Detect whether an autoencoder is robust
}
\seealso{
\verb{\link{correntropy}}, \verb{\link{autoencoder_robust}}, \verb{\link{make_robust}}
}
