% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_sparse.R
\name{is_sparse}
\alias{is_sparse}
\title{Detect whether an autoencoder is sparse}
\usage{
is_sparse(learner)
}
\arguments{
\item{learner}{A \code{"ruta_autoencoder"} object}
}
\value{
Logical value indicating if a sparsity regularization in the encoding layer was found
}
\description{
Detect whether an autoencoder is sparse
}
\seealso{
\code{\link{sparsity}}, \code{\link{autoencoder_sparse}}, \code{\link{make_sparse}}
}
