% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{is_trained}
\alias{is_trained}
\title{Detect trained models}
\usage{
is_trained(learner)
}
\arguments{
\item{learner}{Learner object}
}
\value{
A boolean
}
\description{
Inspects a learner and figures out whether it has been trained
}
\seealso{
\code{\link{train}}
}
