% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_variational.R
\name{is_variational}
\alias{is_variational}
\title{Detect whether an autoencoder is variational}
\usage{
is_variational(learner)
}
\arguments{
\item{learner}{A \code{"ruta_autoencoder"} object}
}
\value{
Logical value indicating if a variational loss was found
}
\description{
Detect whether an autoencoder is variational
}
\seealso{
\code{\link{autoencoder_variational}}
}
