% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.r
\name{+.ruta_network}
\alias{+.ruta_network}
\alias{c.ruta_network}
\title{Add layers to a network/Join networks}
\usage{
\method{+}{ruta_network}(e1, e2)

\method{c}{ruta_network}(...)
}
\arguments{
\item{e1}{First network}

\item{e2}{Second network}

\item{...}{networks or layers to be concatenated}
}
\value{
Network combination
}
\description{
Add layers to a network/Join networks
}
\examples{
network = input() + dense(30) + output("sigmoid")
another = c(input(), dense(30), dense(3), dense(30), output())
}
