% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layer_keras}
\alias{layer_keras}
\title{Custom layer from Keras}
\usage{
layer_keras(name, ...)
}
\arguments{
\item{name}{The name of the layer, e.g. \code{"activity_regularization"} for a
\code{keras::layer_activity_regularization} object.}
}
\value{
A wrapper for the specified layer, which can be combined with other Ruta
layers
}
\description{
Gets any layer available in Keras with the specified parameters
}
\seealso{
Other neural layers: \code{\link{dense}},
  \code{\link{dropout}}, \code{\link{input}},
  \code{\link{output}}, \code{\link{variational_block}}
}
