% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_contractive.R
\name{make_contractive}
\alias{make_contractive}
\title{Add contractive behavior to any autoencoder}
\usage{
make_contractive(learner, weight = 2e-04)
}
\arguments{
\item{learner}{The \code{"ruta_autoencoder"} object}

\item{weight}{Weight assigned to the contractive loss}
}
\value{
An autoencoder object which contains the contractive loss
}
\description{
Converts an autoencoder into a contractive one by assigning a
contractive loss to it
}
\seealso{
\code{\link{autoencoder_contractive}}
}
