% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_denoising.R
\name{make_denoising}
\alias{make_denoising}
\title{Add denoising behavior to any autoencoder}
\usage{
make_denoising(learner, ratio = 0.05)
}
\arguments{
\item{learner}{The \code{"ruta_autoencoder"} object}

\item{ratio}{The ratio of each instance in the input data which will be
altered by random noise}
}
\value{
An autoencoder object which contains the noisy filter
}
\description{
Converts an autoencoder into a denoising one by adding a filter
for the input data
}
