% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_sparse.R
\name{make_sparse}
\alias{make_sparse}
\title{Add sparsity regularization to an autoencoder}
\usage{
make_sparse(learner, high_probability = 0.1, weight = 0.2)
}
\description{
Add sparsity regularization to an autoencoder
}
\seealso{
\code{\link{sparsity}}, \code{\link{autoencoder_sparse}}
}
