% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.r
\name{new_autoencoder}
\alias{new_autoencoder}
\title{Create an autoencoder learner}
\usage{
new_autoencoder(network, loss)
}
\arguments{
\item{network}{Layer construct of class \code{"ruta_network"}}

\item{loss}{Character string specifying a loss function}
}
\value{
A construct of class \code{"ruta_autoencoder"}
}
\description{
Internal function to create autoencoder objects. Instead, consider using
\code{\link{autoencoder}}.
}
