% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{noise}
\alias{noise}
\title{Noise generator}
\usage{
noise(type, ...)
}
\arguments{
\item{type}{Type of noise, as a character string}
}
\description{
Delegates on noise classes to generate noise of some type
}
