% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{noise_cauchy}
\alias{noise_cauchy}
\title{Additive Cauchy noise}
\usage{
noise_cauchy(scale = 0.005)
}
\arguments{
\item{scale}{Scale for the Cauchy distribution}
}
\value{
Object which can be applied to data with \code{\link{apply_filter}}
}
\description{
A data filter which adds noise from a Cauchy distribution to instances
}
\seealso{
Other noise generators: \code{\link{noise_gaussian}},
  \code{\link{noise_ones}}, \code{\link{noise_saltpepper}},
  \code{\link{noise_zeros}}
}
