% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{noise_gaussian}
\alias{noise_gaussian}
\title{Additive Gaussian noise}
\usage{
noise_gaussian(sd = NULL, var = NULL)
}
\arguments{
\item{sd}{Standard deviation for the Gaussian distribution}

\item{var}{Variance of the Gaussian distribution (optional, only used
if \code{sd} is not provided)}
}
\value{
Object which can be applied to data with \code{\link{apply_filter}}
}
\description{
A data filter which adds Gaussian noise to instances
}
\seealso{
Other noise generators: \code{\link{noise_cauchy}},
  \code{\link{noise_ones}}, \code{\link{noise_saltpepper}},
  \code{\link{noise_zeros}}
}
