% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{noise_ones}
\alias{noise_ones}
\title{Filter to add ones noise}
\usage{
noise_ones(p = 0.05)
}
\arguments{
\item{p}{Probability that a feature in an instance is set to one}
}
\value{
Object which can be applied to data with \code{\link{apply_filter}}
}
\description{
A data filter which replaces some values with ones
}
\seealso{
Other noise generators: \code{\link{noise_cauchy}},
  \code{\link{noise_gaussian}},
  \code{\link{noise_saltpepper}}, \code{\link{noise_zeros}}
}
\concept{noise generators}
