% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{noise_saltpepper}
\alias{noise_saltpepper}
\title{Filter to add salt-and-pepper noise}
\usage{
noise_saltpepper(p = 0.05)
}
\arguments{
\item{p}{Probability that a feature in an instance is set to zero or one}
}
\value{
Object which can be applied to data with \code{\link{apply_filter}}
}
\description{
A data filter which replaces some values with zeros or ones
}
\seealso{
Other noise generators: \code{\link{noise_cauchy}},
  \code{\link{noise_gaussian}}, \code{\link{noise_ones}},
  \code{\link{noise_zeros}}
}
