% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{noise_zeros}
\alias{noise_zeros}
\title{Filter to add zero noise}
\usage{
noise_zeros(p = 0.05)
}
\arguments{
\item{p}{Probability that a feature in an instance is set to zero}
}
\value{
Object which can be applied to data with \code{\link{apply_filter}}
}
\description{
A data filter which replaces some values with zeros
}
\seealso{
Other noise generators: \code{\link{noise_cauchy}},
  \code{\link{noise_gaussian}}, \code{\link{noise_ones}},
  \code{\link{noise_saltpepper}}
}
