% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_orthonormal.R
\name{orthonormal_loss}
\alias{orthonormal_loss}
\title{Contractive loss}
\usage{
orthonormal_loss(reconstruction_loss = "mean_squared_error", weight = 0.001)
}
\arguments{
\item{reconstruction_loss}{Original reconstruction error to be combined with the
contractive loss (e.g. \code{"binary_crossentropy"})}

\item{weight}{Weight assigned to the contractive loss}
}
\value{
A loss object which can be converted into a Keras loss
}
\description{
This is a wrapper for a loss which induces a contraction in the
latent space.
}
\seealso{
\verb{\link{autoencoder_contractive}}

Other loss functions: 
\code{\link{contraction}()},
\code{\link{correntropy}()},
\code{\link{loss_variational}()}
}
\concept{loss functions}
