% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.r
\name{output}
\alias{output}
\title{Create an output layer}
\usage{
output(activation = "linear")
}
\arguments{
\item{activation}{Optional, string indicating activation function (linear by default)}
}
\description{
Create an output layer
}
