% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.r
\name{plot.rutaNetwork}
\alias{plot.rutaNetwork}
\title{Draw a neural network}
\usage{
\method{plot}{rutaNetwork}(x, ...)
}
\arguments{
\item{x}{A \code{"rutaNetwork"} object}
}
\description{
Draw a neural network
}
