% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.r
\name{+.rutaNetwork}
\alias{+.rutaNetwork}
\title{Add layers to a network}
\usage{
\method{+}{rutaNetwork}(e1, e2)
}
\arguments{
\item{...}{Networks to be combined}
}
\value{
Network combination
}
\description{
Add layers to a network
}
\examples{
network = input() + dense(30) + output("sigmoid")
}
