% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_model.R
\name{predict.rutaModel}
\alias{predict.rutaModel}
\title{Predict outputs for trained models and new data.}
\usage{
\method{predict}{rutaModel}(object, ...)
}
\arguments{
\item{object}{A \code{"rutaModel"} object.}

\item{...}{Custom parameters for MXNet prediction function.}
}
\value{
A matrix containing predictions for each instance in the given task.
}
\description{
You should pass a \code{model} argument and a \code{task} argument at least.
The task can be the same input data the model was trained with, or new data
with the same shape.
Other arguments will be passed to the internal prediction function.
}

