% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R, R/loss.R, R/network.R
\name{print.ruta_autoencoder}
\alias{print.ruta_autoencoder}
\alias{print.ruta_loss_named}
\alias{print.ruta_loss}
\alias{print.ruta_network}
\title{Inspect Ruta objects}
\usage{
\method{print}{ruta_autoencoder}(x, ...)

\method{print}{ruta_loss_named}(x, ...)

\method{print}{ruta_loss}(x, ...)

\method{print}{ruta_network}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Unused}
}
\value{
Invisibly returns the same object passed as parameter
}
\description{
Inspect Ruta objects
}
