% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{print.ruta_network}
\alias{print.ruta_network}
\title{Inspect a neural network}
\usage{
\method{print}{ruta_network}(x, ...)
}
\arguments{
\item{x}{A \code{"ruta_network"} object}

\item{...}{Additional parameters, currently ignored}
}
\value{
Invisibly returns the same object passed as parameter
}
\description{
Inspect a neural network
}
