% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_learner.R
\name{print.rutaAutoencoder}
\alias{print.rutaAutoencoder}
\title{\code{print} method for autoencoders.}
\usage{
\method{print}{rutaAutoencoder}(x, ...)
}
\arguments{
\item{x}{\code{"rutaAutoencoder"} object}

\item{...}{Ignored}
}
\description{
\code{print} method for autoencoders.
}

