% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.r
\name{print.rutaNetwork}
\alias{print.rutaNetwork}
\title{Inspect a neural network}
\usage{
\method{print}{rutaNetwork}(x, ...)
}
\arguments{
\item{x}{A \code{"rutaNetwork"} object}

\item{...}{Additional parameters, currently ignored}
}
\description{
Inspect a neural network
}
