% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbm_learner.R
\name{print.rutaRBM}
\alias{print.rutaRBM}
\title{\code{print} method for RBMs.}
\usage{
\method{print}{rutaRBM}(x, ...)
}
\arguments{
\item{x}{\code{"rutaRBM"} object}

\item{...}{Ignored}
}
\description{
\code{print} method for RBMs.
}

