% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{print.rutaTask}
\alias{print.rutaTask}
\title{\code{print} method for tasks.}
\usage{
\method{print}{rutaTask}(x, ...)
}
\arguments{
\item{x}{\code{"rutaTask"} object}

\item{...}{Ignored}
}
\description{
\code{print} method for tasks.
}

