% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{reconstruct}
\alias{reconstruct}
\alias{predict.ruta_autoencoder}
\title{Retrieve reconstructions for input data}
\usage{
reconstruct(learner, data)

\method{predict}{ruta_autoencoder}(object, ...)
}
\arguments{
\item{learner}{Trained autoencoder model}

\item{data}{data.frame to be passed through the network}

\item{object}{Trained autoencoder model}

\item{...}{Rest of parameters, unused}
}
\value{
Matrix containing the reconstructions
}
\description{
Extracts the reconstructions calculated by a trained autoencoder for the specified
input data after encoding and decoding. \code{predict} is an alias for \code{reconstruct}.
}
\seealso{
\code{\link{encode}}, \code{\link{decode}}
}
