% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_model.R
\name{ruta.deepFeatures}
\alias{ruta.deepFeatures}
\title{Get the deep features from a trained autoencoder model.}
\usage{
ruta.deepFeatures(model, task, ...)
}
\arguments{
\item{model}{A \code{"rutaModel"} object from an autoencoder learner.}

\item{task}{A \code{"rutaTask"} object.}

\item{...}{Custom parameters for internal prediction function.}
}
\value{
A matrix containing the encoding output for each instance in the
  given task.
}
\description{
Get the deep features from a trained autoencoder model.
}

