% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{ruta.getWeights}
\alias{ruta.getWeights}
\title{Get weights from any layer from a trained model.}
\usage{
ruta.getWeights(model, layer)
}
\arguments{
\item{model}{A \code{"rutaModel"} object.}

\item{layer}{An integer indicating the index of the layer for weights to be
extracted from.}
}
\value{
A matrix containing layer weights.
}
\description{
Get weights from any layer from a trained model.
}

