% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbm_model.R
\name{ruta.getWeights.rutaRBMModel}
\alias{ruta.getWeights.rutaRBMModel}
\title{Get weights from any layer from a trained model.}
\usage{
\method{ruta.getWeights}{rutaRBMModel}(model, layer)
}
\arguments{
\item{model}{A \code{"rutaRBMModel"} object.}

\item{layer}{An integer indicating the index of the layer for weights to be
extracted from.}
}
\value{
A matrix containing layer weights.
}
\description{
Get weights from any layer from a trained model.
}

