% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_model.R
\name{ruta.layerOutputs}
\alias{ruta.layerOutputs}
\title{Get outputs from any layer from a trained model and new data.}
\usage{
ruta.layerOutputs(model, task, layerInput = 1, layerOutput, ...)
}
\arguments{
\item{model}{A \code{"rutaModel"} object from an autoencoder learner.}

\item{task}{A \code{"rutaTask"} object.}

\item{layerInput}{Currently input is always injected into the first layer.}

\item{layerOutput}{An integer indicating the index of the layer to be
obtained.}

\item{...}{Custom parameters for internal prediction function.}
}
\value{
A matrix containing layer outputs for each instance in the given task.
}
\description{
Get outputs from any layer from a trained model and new data.
}

