% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbm_model.R
\name{ruta.layerOutputs.rutaRBMModel}
\alias{ruta.layerOutputs.rutaRBMModel}
\title{Get outputs from any layer from a trained model and new data.}
\usage{
\method{ruta.layerOutputs}{rutaRBMModel}(model, task, layerInput = 1,
  layerOutput = 2 - layerInput, ...)
}
\arguments{
\item{model}{A \code{"rutaRBMModel"} object trained with an RBM learner.}

\item{task}{A \code{"rutaTask"} object.}

\item{layerInput}{Just 1 or 2 (default = 1).}

\item{layerOutput}{The remaining layer (optional).}

\item{...}{Custom parameters for internal prediction function.}
}
\value{
A matrix containing layer outputs for each instance in the given task.
}
\description{
Get outputs from any layer from a trained model and new data.
}

