% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learner.R
\name{ruta.makeLearner}
\alias{ruta.makeLearner}
\title{Create a representation for a learning algorithm.}
\usage{
ruta.makeLearner(cl, id = cl, ...)
}
\arguments{
\item{cl}{A string. Type of learner.}

\item{id}{A string. A name for the learner.}

\item{...}{Additional parameters for the learner.}
}
\value{
An object for the learner containing the provided parameters.
}
\description{
Create a representation for a learning algorithm.
}
\examples{
\dontrun{
ruta.makeLearner("autoencoder", "ae1", hidden = c(4, 2, 4), activation = "relu")
}
}

