% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{ruta.makeTask}
\alias{ruta.makeTask}
\title{Create a representation for a learning task.}
\usage{
ruta.makeTask(id, data, cl = NULL)
}
\arguments{
\item{id}{A string. A name for the task.}

\item{data}{A dataset, will be casted to data.frame with \code{as.data.frame}.}

\item{cl}{The index of the column corresponding to the class, or \code{NULL} if the dataset has no class.}
}
\value{
A generic task object with the provided data and parameters.
}
\description{
Create a representation for a learning task.
}

