% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{[.ruta_network}
\alias{[.ruta_network}
\title{Access subnetworks of a network}
\usage{
\method{[}{ruta_network}(net, index)
}
\arguments{
\item{net}{A \code{"ruta_network"} object}

\item{index}{An integer vector of indices of layers to be extracted}
}
\value{
A \code{"ruta_network"} object containing the specified layers.
}
\description{
Access subnetworks of a network
}
\examples{
(input() + dense(30))[2]
long <- input() + dense(1000) + dense(100) + dense(1000) + output()
short <- long[c(1, 3, 5)]
}
