% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.r
\name{[.rutaNetwork}
\alias{[.rutaNetwork}
\title{Access subnetworks of a network}
\usage{
\method{[}{rutaNetwork}(net, index)
}
\arguments{
\item{net}{A \code{"rutaNetwork"} object}

\item{index}{An integer vector of indices of layers to be extracted}
}
\value{
A \code{"rutaNetwork"} object containing the specified layers.
}
\description{
Access subnetworks of a network
}
\examples{
(input() + dense(30))[2]
long = input() + dense(1000) + dense(100) + dense(1000) + output()
short = long[c(1, 3, 5)]
}
