% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{to_keras}
\alias{to_keras}
\title{Convert a Ruta object onto Keras objects and functions}
\usage{
to_keras(x, ...)
}
\arguments{
\item{x}{Object to be converted}

\item{...}{Remaining parameters depending on the method}
}
\description{
Generic function which uses the Keras API to build objects
out of Ruta wrappers
}
