% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_robust.R
\name{to_keras.ruta_correntropy}
\alias{to_keras.ruta_correntropy}
\title{Obtain a Keras correntropy loss}
\usage{
\method{to_keras}{ruta_correntropy}(x, ...)
}
\arguments{
\item{x}{A \code{"ruta_correntropy"} object}

\item{...}{Rest of parameters, ignored}
}
\description{
Builds the Keras loss function corresponding to the object received
}
