% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{to_keras.ruta_filter}
\alias{to_keras.ruta_filter}
\title{Get a Keras generator from a data filter}
\usage{
\method{to_keras}{ruta_filter}(x, data, batch_size, ...)
}
\arguments{
\item{x}{Filter object}

\item{data}{Matrix where the filter will be applied}

\item{batch_size}{Size of the sample (for the training stage)}

\item{...}{Additional parameters, currently unused}
}
\description{
Noise filters can be applied during training (in denoising autoencoders),
for this a generator is used to get data batches.
}
