% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{to_keras.ruta_layer_input}
\alias{to_keras.ruta_layer_input}
\alias{to_keras.ruta_layer_dense}
\alias{to_keras.ruta_layer_conv}
\alias{to_keras.ruta_layer_custom}
\title{Convert Ruta layers onto Keras layers}
\usage{
\method{to_keras}{ruta_layer_input}(x, input_shape, ...)

\method{to_keras}{ruta_layer_dense}(x, input_shape,
  model = keras::keras_model_sequential(), ...)

\method{to_keras}{ruta_layer_conv}(x, input_shape,
  model = keras::keras_model_sequential(), ...)

\method{to_keras}{ruta_layer_custom}(x, input_shape,
  model = keras::keras_model_sequential(), ...)
}
\arguments{
\item{x}{The layer object}

\item{input_shape}{Number of features in training data}

\item{...}{Unused}

\item{model}{Keras model where the layer will be added}
}
\value{
A Layer object from Keras
}
\description{
Convert Ruta layers onto Keras layers
}
