% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_contractive.R
\name{to_keras.ruta_loss_contraction}
\alias{to_keras.ruta_loss_contraction}
\title{Obtain a Keras contractive loss}
\usage{
\method{to_keras}{ruta_loss_contraction}(x, keras_model, ...)
}
\arguments{
\item{x}{A \code{"ruta_loss_contraction"} object}

\item{keras_model}{The keras autoencoder which will use the loss function}

\item{...}{Rest of parameters, ignored}
}
\value{
A function which returns the contractive loss for given true and
predicted values
}
\description{
Builds the Keras loss function corresponding to the object received
}
\references{
\href{https://wiseodd.github.io/techblog/2016/12/05/contractive-autoencoder/}{Deriving Contractive Autoencoder and Implementing it in Keras}
}
\seealso{
\code{\link{autoencoder_contractive}}
}
