% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_variational.R
\name{to_keras.ruta_loss_variational}
\alias{to_keras.ruta_loss_variational}
\title{Obtain a Keras variational loss}
\usage{
\method{to_keras}{ruta_loss_variational}(loss, keras_model, ...)
}
\arguments{
\item{keras_model}{The keras autoencoder which will use the loss function}

\item{...}{Rest of parameters, ignored}

\item{x}{A \code{"ruta_loss_variational"} object}
}
\value{
A function which returns the variational loss for given true and
predicted values
}
\description{
Builds the Keras loss function corresponding to the object received
}
\references{
\itemize{
\item \href{https://arxiv.org/abs/1312.6114}{Auto-Encoding Variational Bayes}
\item \href{http://blog.fastforwardlabs.com/2016/08/22/under-the-hood-of-the-variational-autoencoder-in.html}{Under the Hood of the Variational Autoencoder (in Prose and Code)}
\item \href{https://keras.rstudio.com/articles/examples/variational_autoencoder.html}{Keras example: Variational autoencoder}
}
}
\seealso{
\code{\link{autoencoder_variational}}
}
