% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.r
\name{to_keras.ruta_network}
\alias{to_keras.ruta_network}
\title{Build a Keras network}
\usage{
\method{to_keras}{ruta_network}(x, input_shape)
}
\arguments{
\item{x}{A \code{"ruta_network"} object}

\item{input_shape}{The length of each input vector (number of input attributes)}
}
\value{
A list of Keras Tensor objects with an attribute \code{"encoding"}
indicating the index of the encoding layer
}
\description{
Build a Keras network
}
