% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_sparse.R
\name{to_keras.ruta_sparsity}
\alias{to_keras.ruta_sparsity}
\title{Translate sparsity regularization to Keras regularizer}
\usage{
\method{to_keras}{ruta_sparsity}(x, activation)
}
\arguments{
\item{x}{Sparsity object}

\item{activation}{Name of the activation function used in the encoding layer}
}
\value{
Function which can be used as activity regularizer in a Keras layer
}
\description{
Translate sparsity regularization to Keras regularizer
}
\references{
Andrew Ng, Sparse Autoencoder.
\href{https://web.stanford.edu/class/cs294a/sparseAutoencoder_2011new.pdf}{CS294A Lecture Notes} (2011)
}
