% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_weight_decay.R
\name{to_keras.ruta_weight_decay}
\alias{to_keras.ruta_weight_decay}
\title{Obtain a Keras weight decay}
\usage{
\method{to_keras}{ruta_weight_decay}(x, ...)
}
\arguments{
\item{x}{A \code{"ruta_weight_decay"} object}

\item{...}{Rest of parameters, ignored}
}
\description{
Builds the Keras regularizer corresponding to the weight decay
}
