% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.r
\name{toKeras}
\alias{toKeras}
\title{Build a Keras network}
\usage{
toKeras(net, input_shape)
}
\arguments{
\item{net}{A \code{"rutaNetwork"} object}

\item{input_shape}{The length of each input vector (number of input attributes)}
}
\value{
A list of Keras Tensor objects with an attribute \code{"encoding"}
indicating the index of the encoding layer
}
\description{
Build a Keras network
}
