% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{train}
\alias{train}
\title{Train a learner object.}
\usage{
train(x, ...)
}
\arguments{
\item{x}{Learner (possibly a \code{"rutaLearner"} object).}

\item{...}{Parameters passed to the specific method.}
}
\value{
A trained model.
}
\description{
\code{train} creates a trained model for a given learner.
}
\details{
This is a generic function, will find a specific method for each learner class.
}

