% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_model.R
\name{train.rutaAutoencoder}
\alias{train.rutaAutoencoder}
\title{Train an autoencoder.}
\usage{
\method{train}{rutaAutoencoder}(x, task, ...)
}
\arguments{
\item{x}{A \code{"rutaAutoencoder"} object. This is the learner that will be
trained.}

\item{task}{A \code{"rutaUnsupervisedTask"} object. It contains the data that
will be used to train the autoencoder. Class information, if present, will
not be used.}

\item{...}{Additional parameters for the MXNet optimizer}
}
\value{
A \code{"rutaModel"} object containing the trained model.
}
\description{
Train an autoencoder.
}

