% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbm_model.R
\name{train.rutaRBM}
\alias{train.rutaRBM}
\title{Train a Restricted Boltzmann Machine.}
\usage{
\method{train}{rutaRBM}(x, task, ...)
}
\arguments{
\item{x}{A \code{"rutaRBM"} object. This is the learner that will be
trained.}

\item{task}{A \code{"rutaUnsupervisedTask"} object. It contains the data that
will be used to train the RBM. Class information, if present, will
not be used.}

\item{...}{Additional parameters for the training function}
}
\value{
A \code{"rutaModel"} object containing the trained model.
}
\description{
Train a Restricted Boltzmann Machine.
}

