% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_variational.R
\name{variational_loss}
\alias{variational_loss}
\title{Variational loss}
\usage{
variational_loss(base_loss)
}
\arguments{
\item{base_loss}{Another loss to be used as reconstruction error (e.g. "binary_crossentropy")}
}
\value{
A \code{"ruta_loss"} object
}
\description{
Specifies an evaluation function adapted to the variational autoencoder. It combines
a base reconstruction error and the Kullback-Leibler divergence between the learned
distribution and the true latent posterior.
}
