% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_weight_decay.R
\name{weight_decay}
\alias{weight_decay}
\title{Weight decay}
\usage{
weight_decay(decay = 0.02)
}
\arguments{
\item{decay}{Numeric value indicating the amount of decay}
}
\value{
A regularizer object containing the set parameters
}
\description{
A wrapper that describes a weight decay regularization of the
encoding layer
}
