% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.r
\name{autoencoder}
\alias{autoencoder}
\title{Create an autoencoder learner}
\usage{
autoencoder(layers, loss = "mse")
}
\arguments{
\item{layers}{Layer construct of class \code{"rutaNetwork"}}

\item{loss}{Character string specifying a loss function}
}
\description{
Create an autoencoder learner
}
