% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{autoencoder}
\alias{autoencoder}
\title{Create an autoencoder learner}
\usage{
autoencoder(network, loss = "mse")
}
\arguments{
\item{network}{Layer construct of class \code{"ruta_network"}}

\item{loss}{Character string specifying a loss function}
}
\value{
A construct of class \code{"ruta_autoencoder"}
}
\description{
Represents a generic autoencoder network.
}
\seealso{
\code{\link{train.ruta_autoencoder}}
}
