% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.r
\name{autoencoder}
\alias{autoencoder}
\title{Create an autoencoder learner}
\usage{
autoencoder(layers, sparse = F, contractive = F)
}
\arguments{
\item{contractive}{}
}
\description{
Create an autoencoder learner
}
