% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_variational.R
\name{autoencoder_variational}
\alias{autoencoder_variational}
\title{This file contains excerpts of code from Keras examples demonstrating how to
build a variational autoencoder with Keras.
Source: https://github.com/keras-team/keras/blob/master/examples/variational_autoencoder.py
Reference: "Auto-Encoding Variational Bayes" https://arxiv.org/abs/1312.6114
Build a variational autoencoder}
\usage{
autoencoder_variational(intermediate, latent_dim,
  loss = "loss_binary_crossentropy")
}
\arguments{
\item{intermediate}{Intermediate size}

\item{latent_dim}{Latent space dimension}

\item{loss}{Reconstruction error to be combined with KL divergence in order to compute
the variational loss}
}
\description{
This file contains excerpts of code from Keras examples demonstrating how to
build a variational autoencoder with Keras.
Source: https://github.com/keras-team/keras/blob/master/examples/variational_autoencoder.py
Reference: "Auto-Encoding Variational Bayes" https://arxiv.org/abs/1312.6114
Build a variational autoencoder
}
