% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_contractive.R
\name{contraction}
\alias{contraction}
\title{Contractive loss}
\usage{
contraction(reconstruction_loss = "mean_squared_error", weight = 2e-04)
}
\arguments{
\item{reconstruction_loss}{Original reconstruction error to be combined with the
contractive loss (e.g. \code{"binary_crossentropy"})}

\item{weight}{Weight assigned to the contractive loss}
}
\value{
A loss object which can be converted into a Keras loss
}
\description{
This is a wrapper for a loss which induces a contraction in the
latent space.
}
\seealso{
\code{\link{autoencoder_contractive}}

Other loss functions: \code{\link{correntropy}},
  \code{\link{loss_variational}}
}
\concept{loss functions}
