% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{decode}
\alias{decode}
\title{Retrieve decoding of encoded data}
\usage{
decode(learner, data)
}
\arguments{
\item{learner}{Trained autoencoder model}

\item{data}{data.frame to be decoded}
}
\value{
Matrix containing the decodifications
}
\description{
Extracts the decodification calculated by a trained autoencoder for the specified
data.
}
\seealso{
\code{\link{encode}}, \code{\link{reconstruct}}
}
