% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.r
\name{dense}
\alias{dense}
\title{Create a fully-connected layer}
\usage{
dense(units, activation = "linear")
}
\arguments{
\item{units}{Number of units}

\item{activation}{Optional, string indicating activation function (linear by default)}
}
\description{
Create a fully-connected layer
}
