% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{encode}
\alias{encode}
\title{Retrieve encoding of data}
\usage{
encode(learner, data)
}
\arguments{
\item{learner}{Trained autoencoder model}

\item{data}{data.frame to be encoded}
}
\value{
Matrix containing the encodings
}
\description{
Extracts the encoding calculated by a trained autoencoder for the specified
data.
}
\seealso{
\code{\link{decode}}, \code{\link{reconstruct}}
}
