% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.r
\name{input}
\alias{input}
\title{Create an input layer}
\usage{
input()
}
\value{
A construct with class \code{"ruta_layer"}
}
\description{
This layer acts as a placeholder for input data. The number of units is not
needed as it is deduced from the data during training.
}
\seealso{
Other neural layers: \code{\link{dense}},
  \code{\link{output}}
}
