% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_contractive.R
\name{is_contractive}
\alias{is_contractive}
\title{Detect whether an autoencoder is contractive}
\usage{
is_contractive(learner)
}
\arguments{
\item{learner}{A \code{"ruta_autoencoder"} object}
}
\value{
Logical value indicating if a contractive loss was found
}
\description{
Detect whether an autoencoder is contractive
}
\seealso{
\code{\link{contraction}}, \code{\link{autoencoder_contractive}}, \code{\link{make_contractive}}
}
