% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.r
\name{+.rutaNetwork}
\alias{+.rutaNetwork}
\alias{c.rutaNetwork}
\title{Add layers to a network/Join networks}
\usage{
\method{+}{rutaNetwork}(e1, e2)

\method{c}{rutaNetwork}(...)
}
\arguments{
\item{e1}{First network}

\item{e2}{Second network}

\item{...}{networks to be concatenated}
}
\value{
Network combination
}
\description{
Add layers to a network/Join networks
}
\examples{
network = input() + dense(30) + output("sigmoid")
another = c(input(), dense(30), dense(3), dense(30), output())
}
