% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_denoising.R
\name{make_denoising}
\alias{make_denoising}
\title{Add denoising behavior to any autoencoder}
\usage{
make_denoising(learner, noise_type = "zeros", ...)
}
\arguments{
\item{learner}{The \code{"ruta_autoencoder"} object}

\item{noise_type}{Type of data corruption which will be used to train the
autoencoder, as a character string. See \code{\link{autoencoder_denoising}} for
details}

\item{...}{Extra parameters to customize the noisy filter. See
\code{\link{autoencoder_denoising}} for details}
}
\value{
An autoencoder object which contains the noisy filter
}
\description{
Converts an autoencoder into a denoising one by adding a filter
for the input data
}
\seealso{
\code{\link{autoencoder_denoising}}
}
