% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.R
\name{new_autoencoder}
\alias{new_autoencoder}
\title{Create an autoencoder learner}
\usage{
new_autoencoder(network, loss, extra_class = NULL)
}
\arguments{
\item{network}{Layer construct of class \code{"ruta_network"} or coercible}

\item{loss}{A \code{"ruta_loss"} object or a character string specifying a loss function}

\item{extra_class}{Vector of classes in case this autoencoder needs to support custom
methods (for \code{to_keras}, \code{train}, \code{generate} or others)}
}
\value{
A construct of class \code{"ruta_autoencoder"}
}
\description{
Internal function to create autoencoder objects. Instead, consider using
\code{\link{autoencoder}}.
}
