% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{new_layer}
\alias{new_layer}
\title{Layer wrapper constructor}
\usage{
new_layer(cl, ...)
}
\arguments{
\item{cl}{Character string specifying class of layer (e.g. \code{"ruta_layer_dense"}), which
will be used to call the corresponding methods}

\item{...}{Other parameters (usually \code{units}, \code{activation})}
}
\value{
A construct with class \code{"ruta_layer"}
}
\description{
Constructor function for layers. You shouldn't generally need to use this. Instead, consider
using individual functions such as \code{\link{dense}}.
}
\examples{
my_layer <- new_layer("dense", 30, "tanh")

# Equivalent:
my_layer <- dense(30, "tanh")[[1]]
}
