% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{output}
\alias{output}
\title{Create an output layer}
\usage{
output(activation = "linear")
}
\arguments{
\item{activation}{Optional, string indicating activation function (linear by default)}
}
\value{
A construct with class \code{"ruta_network"}
}
\description{
This layer acts as a placeholder for the output layer in an autoencoder. The
number of units is not needed as it is deduced from the data during training.
}
\seealso{
Other neural layers: \code{\link{conv}},
  \code{\link{dense}}, \code{\link{dropout}},
  \code{\link{input}}, \code{\link{layer_keras}},
  \code{\link{variational_block}}
}
\concept{neural layers}
