% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.r
\name{plot.ruta_network}
\alias{plot.ruta_network}
\title{Draw a neural network}
\usage{
\method{plot}{ruta_network}(x, ...)
}
\arguments{
\item{x}{A \code{"ruta_network"} object}
}
\description{
Draw a neural network
}
