% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder_model.R
\name{ruta.pretrain}
\alias{ruta.pretrain}
\title{Pretrain autoencoders.}
\usage{
ruta.pretrain(x, task, epochs = 10, ...)
}
\description{
Obtain initial weights for an autoencoder via stack of RBMs.
}

