% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoencoder.r
\name{train}
\alias{train}
\alias{train.ruta_autoencoder}
\title{Train a learner object with data}
\usage{
train(learner, ...)

\method{train}{ruta_autoencoder}(learner, data, validation_data = NULL,
  epochs = 100, ...)
}
\arguments{
\item{learner}{A \code{"ruta_autoencoder"} object}

\item{...}{Additional parameters for \code{keras::fit}}

\item{data}{Training data: columns are attributes and rows are instances}

\item{validation_data}{Additional data.frame of data which will not be used
for training but the loss measure will be calculated against it}

\item{epochs}{The number of times data will pass through the network}
}
\value{
Same autoencoder passed as parameter, with trained internal models
}
\description{
This function compiles the neural network described by the learner object
and trains it with the input data.
}
